# src/main/java/io/swagger/api/OrdersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.RetailOrder;
import io.swagger.model.SalesRep;
import io.swagger.model.WholesaleOrder;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-03-03T19:46:44.474Z")

@Api(value = "orders", description = "the orders API")
public interface OrdersApi {

    @ApiOperation(value = "", nickname = "addRetailOrder", notes = "Add a new retail order to the sales system", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created") })
    @RequestMapping(value = "/orders/retail/new",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<RetailOrder> addRetailOrder(@ApiParam(value = "Retail order object that needs to be added to the Sales System" ,required=true )  @Valid @RequestBody RetailOrder body);


    @ApiOperation(value = "", nickname = "addWholesaleOrder", notes = "Add a new wholesale order to the sales system", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created") })
    @RequestMapping(value = "/orders/wholesale/new",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<WholesaleOrder> addWholesaleOrder(@ApiParam(value = "Add a new wholesale order to the sales system" ,required=true )  @Valid @RequestBody WholesaleOrder body);


    @ApiOperation(value = "", nickname = "changeOrderStatus", notes = "Update the status of an order", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 202, message = "Updated") })
    @RequestMapping(value = "/orders/update/status",
        produces = { "application/json" }, 
        method = RequestMethod.PATCH)
    ResponseEntity<RetailOrder> changeOrderStatus(@ApiParam(value = "ID identifying the Order" ,required=true )  @Valid @RequestBody Long id,
                                           @ApiParam(value = "Status to change on the Order" ,required=true )  @Valid @RequestBody RetailOrder.StatusEnum status);


    @ApiOperation(value = "", nickname = "getOrder", notes = "Gets a completed Order", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK") })
    @RequestMapping(value = "/orders/completed",
        produces = { "application/json" }, 
        method = RequestMethod.GET)

    ResponseEntity<RetailOrder> getOrder( @NotNull@ApiParam(value = "", required = true) @RequestParam(value = "serial_num", required = true) String serialNum) throws NotFoundException;

    @ApiOperation(value = "", notes = "Gets all orders for a given sales rep", response = Void.class, tags={  })
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Created", response = Void.class) })

    @RequestMapping(value = "/orders/salesrep",
            produces = { "application/json" },
            method = RequestMethod.GET)
    ResponseEntity<List<WholesaleOrder>> getOrdersByRep(@NotNull@ApiParam(value = "", required = true) @RequestParam(value = "sales_rep_id", required = true) String salesRepId) throws NotFoundException;

    @ApiOperation(value = "", nickname = "getSalesRep", notes = "HR Endpoint for getting information about sales rep", tags={  })
    @ApiResponses(value = {
        @ApiResponse(code = 200, message = "OK") })
    @RequestMapping(value = "/orders",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<SalesRep> getSalesRep(@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "sales_rep_id", required = true) String salesRepId, @NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "date_from", required = true) String dateFrom, @NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "date_to", required = true) String dateTo);


    @ApiOperation(value = "", nickname = "zeroDollarOrder", notes = "Support submitting of $0 orders", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created") })
    @RequestMapping(value = "/orders/new/refund",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<RetailOrder> zeroDollarOrder(@ApiParam(value = "Support submitting of $0 orders" ,required=true )  @Valid @RequestBody RetailOrder body);

}